+++ src/Randomizer/ShortenPrayerPatch.js
@@ -1,0 +1,10 @@
+/* eslint import/no-webpack-loader-syntax: off */
+import patchText from '!array-loader!./tables/shorten_prayer_patch.txt';
+import { utils } from 'randomtools-js';
+
+const ShortenPrayerPatch = {
+    name: "Shorten prayer",
+    entries: utils.standardPatchLoader(patchText),
+};
+
+export default ShortenPrayerPatch;